refactor(hydra): move DocumentationNormalizer Hydra operations methods to HydraOperationsTrait
